// src/app/api/account.ts
export interface Account {
    ID?: number;
    USERNAME?: string;
    PASSWORD?: string;
    FULL_NAME?: string;
    DIVISI?: any;
    EMAIL?: string;
    ACCESS_ROLE?: any;
    STATUS?: string;
    CREATED_AT?: string;
    MODIFIED_AT?: string;
    ID_ROLE?: string;
    ROLE_NAME?: any;
}